requirements.txt
jupyter
jupyterlab
pandas
matplotlib
ipympl
lxml
django-environ
